+++ etlrules/backends/common/io/files.py
@@ -6,25 +6,34 @@ from etlrules.backends.common.substitution import subst_string
 
 
 class BaseReadFileRule(BaseRule):
-    def __init__(self, file_name: str, file_dir: str=".", regex: bool=False, named_output: Optional[str]=None, name: Optional[str]=None, description: Optional[str]=None, strict: bool=True):
+    def __init__(self, file_name: str, file_dir: Optional[str]=None, regex: bool=False, named_output: Optional[str]=None, name: Optional[str]=None, description: Optional[str]=None, strict: bool=True):
         super().__init__(named_output=named_output, name=name, description=description, strict=strict)
         self.file_name = file_name
         self.file_dir = file_dir
         self.regex = bool(regex)
+        if self._is_uri() and self.regex:
+            raise ValueError("Regex read not supported for URIs.")
+
+    def _is_uri(self):
+        file_name = self.file_name.lower()
+        return file_name.startswith("http://") or file_name.startswith("https://")
 
     def has_input(self):
         return False
 
     def _get_full_file_paths(self):
         file_name = subst_string(self.file_name)
-        file_dir = subst_string(self.file_dir)
+        file_dir = subst_string(self.file_dir or "")
         if self.regex:
             pattern = re.compile(file_name)
             for fn in os.listdir(file_dir):
                 if pattern.match(fn):
                     yield os.path.join(file_dir, fn)
         else:
-            yield os.path.join(file_dir, file_name)
+            if self._is_uri():
+                yield file_name
+            else:
+                yield os.path.join(file_dir, file_name)
 
     def do_read(self, file_path: str):
         raise NotImplementedError("Have you imported the rules from etlrules.backends.<your_backend> and not common?")
@@ -73,10 +82,11 @@ class ReadCSVFileRule(BaseReadFileRule):
             all those csv file and concatenate them into a single dataframe.
             For example, file_name=".*\.csv", file_dir=".", regex=True will extract all the files with the .csv extension
             from the current directory.
+            It can also be an URI (e.g. https://example.com/mycsv.csv)
         file_dir: The file directory where the file_name is located. When file_name is a regular expression and 
             the regex parameter is True, file_dir is the directory that is inspected for any files that match the
-            regular expression.
-            Defaults to . (ie the current directory).
+            regular expression. Optional.
+            For files it defaults to . (ie the current directory). Ignored for URIs.
         regex: When True, the file_name is interpreted as a regular expression. Defaults to False.
         separator: The single character to be used as separator in the csv file. Defaults to , (comma).
         header: When True, the first line is interpreted as the header and the column names are extracted from it.
@@ -97,7 +107,7 @@ class ReadCSVFileRule(BaseReadFileRule):
         IOError: raised when the file is not found.
     """
 
-    def __init__(self, file_name: str, file_dir: str=".", regex: bool=False, separator: str=",",
+    def __init__(self, file_name: str, file_dir: Optional[str]=None, regex: bool=False, separator: str=",",
                  header: bool=True, skip_header_rows: Optional[int]=None,
                  named_output: Optional[str]=None, name: Optional[str]=None, description: Optional[str]=None, strict: bool=True):
         super().__init__(file_name=file_name, file_dir=file_dir, regex=regex, named_output=named_output, name=name, description=description, strict=strict)
@@ -297,7 +307,7 @@ class WriteCSVFileRule(BaseWriteFileRule):
         'xz': '.xz',
     }
 
-    def __init__(self, file_name, file_dir=".", separator=",", header=True, compression=None, named_input=None, name=None, description=None, strict=True):
+    def __init__(self, file_name: str, file_dir: str=".", separator: str=",", header: bool=True, compression: Optional[str]=None, named_input: Optional[str]=None, name: Optional[str]=None, description: Optional[str]=None, strict: bool=True):
         super().__init__(
             file_name=file_name, file_dir=file_dir, named_input=named_input, 
             name=name, description=description, strict=strict)